Trait pallet_starknet_runtime_api::StarknetRuntimeApi

pub trait StarknetRuntimeApi<Block: BlockT>: Core<Block> {

    // Provided methods
    fn nonce(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        contract_address: ContractAddress
    ) -> Result<Nonce, ApiError> { ... }
    fn get_storage_at(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        address: ContractAddress,
        key: StorageKey
    ) -> Result<Result<StarkFelt, DispatchError>, ApiError> { ... }
    fn call(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        address: ContractAddress,
        function_selector: EntryPointSelector,
        calldata: Calldata
    ) -> Result<Result<Vec<Felt252Wrapper>, DispatchError>, ApiError> { ... }
    fn contract_class_hash_by_address(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        address: ContractAddress
    ) -> Result<ClassHash, ApiError> { ... }
    fn contract_class_by_class_hash(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        class_hash: ClassHash
    ) -> Result<Option<ContractClass>, ApiError> { ... }
    fn chain_id(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Felt252Wrapper, ApiError> { ... }
    fn program_hash(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Felt252Wrapper, ApiError> { ... }
    fn config_hash(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<StarkHash, ApiError> { ... }
    fn fee_token_address(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<ContractAddress, ApiError> { ... }
    fn estimate_fee(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        transactions: Vec<UserTransaction>
    ) -> Result<Result<Vec<(u64, u64)>, DispatchError>, ApiError> { ... }
    fn estimate_message_fee(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        message: HandleL1MessageTransaction
    ) -> Result<Result<(u128, u64, u64), DispatchError>, ApiError> { ... }
    fn simulate_transactions(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        transactions: Vec<UserTransaction>,
        simulation_flags: SimulationFlags
    ) -> Result<Result<Vec<Result<TransactionExecutionInfo, PlaceHolderErrorTypeForFailedStarknetExecution>>, DispatchError>, ApiError> { ... }
    fn simulate_message(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        message: HandleL1MessageTransaction,
        simulation_flags: SimulationFlags
    ) -> Result<Result<Result<TransactionExecutionInfo, PlaceHolderErrorTypeForFailedStarknetExecution>, DispatchError>, ApiError> { ... }
    fn extrinsic_filter(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        xts: Vec<<Block as BlockT>::Extrinsic>
    ) -> Result<Vec<Transaction>, ApiError> { ... }
    fn re_execute_transactions(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        transactions: Vec<UserOrL1HandlerTransaction>
    ) -> Result<Result<Result<Vec<TransactionExecutionInfo>, PlaceHolderErrorTypeForFailedStarknetExecution>, DispatchError>, ApiError> { ... }
    fn get_index_and_tx_for_tx_hash(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        xts: Vec<<Block as BlockT>::Extrinsic>,
        chain_id: Felt252Wrapper,
        tx_hash: Felt252Wrapper
    ) -> Result<Option<(u32, Transaction)>, ApiError> { ... }
    fn get_events_for_tx_by_hash(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        tx_hash: TransactionHash
    ) -> Result<Vec<StarknetEvent>, ApiError> { ... }
    fn get_tx_execution_outcome(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        tx_hash: TransactionHash
    ) -> Result<Option<Vec<u8>>, ApiError> { ... }
    fn get_block_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<BlockContext, ApiError> { ... }
    fn is_transaction_fee_disabled(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<bool, ApiError> { ... }
    fn get_tx_messages_to_l1(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        tx_hash: TransactionHash
    ) -> Result<Vec<MessageToL1>, ApiError> { ... }
    fn l1_nonce_unused(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        nonce: Nonce
    ) -> Result<bool, ApiError> { ... }

}

Provided Methods

fn nonce( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, contract_address: ContractAddress ) -> Result<Nonce, ApiError>

Returns the nonce associated with the given address in the given block

fn get_storage_at( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, address: ContractAddress, key: StorageKey ) -> Result<Result<StarkFelt, DispatchError>, ApiError>

Returns a storage slot value

fn call( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, address: ContractAddress, function_selector: EntryPointSelector, calldata: Calldata ) -> Result<Result<Vec<Felt252Wrapper>, DispatchError>, ApiError>

Returns a Call response.

fn contract_class_hash_by_address( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, address: ContractAddress ) -> Result<ClassHash, ApiError>

Returns the contract class hash at the given address.

fn contract_class_by_class_hash( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, class_hash: ClassHash ) -> Result<Option<ContractClass>, ApiError>

Returns the contract class for the given class hash.

fn chain_id( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Felt252Wrapper, ApiError>

Returns the chain id.

fn program_hash( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Felt252Wrapper, ApiError>

Returns the Starknet OS Cairo program hash.

fn config_hash( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<StarkHash, ApiError>

Returns the Starknet config hash.

fn fee_token_address( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<ContractAddress, ApiError>

Returns the fee token address.

fn estimate_fee( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, transactions: Vec<UserTransaction> ) -> Result<Result<Vec<(u64, u64)>, DispatchError>, ApiError>

Returns fee estimate

fn estimate_message_fee( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, message: HandleL1MessageTransaction ) -> Result<Result<(u128, u64, u64), DispatchError>, ApiError>

Returns message fee estimate

fn simulate_transactions( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, transactions: Vec<UserTransaction>, simulation_flags: SimulationFlags ) -> Result<Result<Vec<Result<TransactionExecutionInfo, PlaceHolderErrorTypeForFailedStarknetExecution>>, DispatchError>, ApiError>

Simulates transactions and returns their trace

fn simulate_message( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, message: HandleL1MessageTransaction, simulation_flags: SimulationFlags ) -> Result<Result<Result<TransactionExecutionInfo, PlaceHolderErrorTypeForFailedStarknetExecution>, DispatchError>, ApiError>

Simulates single L1 Message and returns its trace

fn extrinsic_filter( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, xts: Vec<<Block as BlockT>::Extrinsic> ) -> Result<Vec<Transaction>, ApiError>

Filters extrinsic transactions to return only Starknet transactions

To support runtime upgrades, the client must be unaware of the specific extrinsic details. To achieve this, the client uses an OpaqueExtrinsic type to represent and manipulate extrinsics. However, the client cannot decode and filter extrinsics due to this limitation. The solution is to offload decoding and filtering to the RuntimeApi in the runtime itself, accomplished through the extrinsic_filter method. This enables the client to operate seamlessly while abstracting the extrinsic complexity.

fn re_execute_transactions( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, transactions: Vec<UserOrL1HandlerTransaction> ) -> Result<Result<Result<Vec<TransactionExecutionInfo>, PlaceHolderErrorTypeForFailedStarknetExecution>, DispatchError>, ApiError>

Re-execute a block and return the TransactionExecutionInfos of every transaction in it, in the same order

fn get_index_and_tx_for_tx_hash( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, xts: Vec<<Block as BlockT>::Extrinsic>, chain_id: Felt252Wrapper, tx_hash: Felt252Wrapper ) -> Result<Option<(u32, Transaction)>, ApiError>

fn get_events_for_tx_by_hash( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, tx_hash: TransactionHash ) -> Result<Vec<StarknetEvent>, ApiError>

fn get_tx_execution_outcome( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, tx_hash: TransactionHash ) -> Result<Option<Vec<u8>>, ApiError>

Return the outcome of the tx execution

fn get_block_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<BlockContext, ApiError>

Return the block context

fn is_transaction_fee_disabled( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<bool, ApiError>

Return is fee disabled in state

fn get_tx_messages_to_l1( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, tx_hash: TransactionHash ) -> Result<Vec<MessageToL1>, ApiError>

Return messages sent to L1 during tx execution

fn l1_nonce_unused( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, nonce: Nonce ) -> Result<bool, ApiError>

Check if L1 Message Nonce has not been used